Initial Linux Support
I redid the report descriptor based on a HORI Fighting Commander 4's
descriptor to get the joystick working with SDL in Linux (mostly to get
it working with RetroPie). It seems that the main issue was having a
dpad (hat) but no actual joystick. This version adds 2 ghost joysticks
and leaves them at 0x7F.