Analog stick support
You can compile to use the joystick as either the left analog stick or
the DPAD. Future version will add support for the right stick and mode
switching on the fly. Joystick now works with the left stick in Big
Picture Mode for Steam. FINALLY works with Shovel Knight. Super Meat Boy
support is broken.